requirements.txt
hydra-core==1.3.2
numpy==1.25.1
omegaconf==2.3.0
Pillow==10.0.0
pydicom==2.4.1
pytorch_lightning==2.0.5
scikit-image==0.21.0
transformers==4.30.2
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
torchvision==0.14.1+cu117
wandb==0.15.5